requirements.txt
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
bcrypt==4.0.1
click==8.1.7
dill==0.3.7
exceptiongroup==1.1.3
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
platformdirs==3.11.0
psycopg2-binary==2.9.9
pydantic==1.10.13
pydantic_core==2.10.1
PyJWT==2.8.0
pylint==3.0.2
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.50
sqlalchemy2-stubs==0.0.2a36
sqlmodel==0.0.11
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.2
typing_extensions==4.8.0
uvicorn==0.23.2
